Address Ranges and Port Ranges
Add the ability to use 'x' to scan a range of 1-256 on that octet ie. (192.168.1.x)
Add the ability to use 'x' and to use ranges in the ports to scan ie. (x, 22,45-80,443)
Adjust the look of the main menu.
Added an offline prompt when connection is interrupted while scanning.
Polished some behavior/messages with prompts.
Added some basic settings. More to come..
Also to come, reports and report formats.